requirement.txt
absl-py==0.1.11
adium-theme-ubuntu==0.3.4
astor==0.6.2
backports.weakref==1.0.post1
BeautifulSoup==3.2.1
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
cssselect==1.0.1
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.7.1
ecdsa==0.13
enum34==1.1.6
feedparser==5.2.1
Flask==0.12.2
ftfy==4.4.3
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
goose-extractor==1.0.25
grpcio==1.10.0
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
jieba==0.39
Jinja2==2.9.6
lxml==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
mock==2.0.0
msgpack-numpy==0.4.2
msgpack-python==0.4.8
murmurhash==0.28.0
nltk==3.2.5
numpy==1.13.3
olefile==0.44
paramiko==1.16.0
pathlib==1.0.1
pbr==3.1.1
Pillow==4.3.0
plac==0.9.6
preshed==1.0.0
protobuf==3.5.2
py4j==0.10.6
pycrypto==2.6.1
PyMySQL==0.7.11
pyspark==2.3.0
regex==2017.4.5
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
spacy==2.0.3
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
tflearn==0.3.2
thinc==6.10.1
toolz==0.8.2
tqdm==4.19.4
ujson==1.35
unity-lens-photos==1.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wrapt==1.10.11
